refactor: move API classes to v1 package for better versioning and organization
