--- src/main/java/org/thingai/app/api/MatchControlApi.java
@@ -1,10 +1,0 @@
-package org.thingai.app.api;
-
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/scorekeeper")
-public class MatchControlApi {
-
-
-}
