--- src/main/java/org/thingai/app/api/TeamApi.java
@@ -1,8 +1,0 @@
-package org.thingai.app.api;
-
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/team")
-public class TeamApi {
-}
