+++ src/main/java/org/thingai/app/api/v1/EventApi.java
@@ -1,10 +1,10 @@
-package org.thingai.app.api;
+package org.thingai.app.api.v1;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/sync")
-public class SyncApi {
+@RequestMapping("/api/event")
+public class EventApi {
 
 }
