+++ src/main/java/org/thingai/app/api/v1/MatchControlApi.java
@@ -1,10 +1,11 @@
-package org.thingai.app.api;
+package org.thingai.app.api.v1;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/rank")
-public class RankApi {
+@RequestMapping("/api/scorekeeper")
+public class MatchControlApi {
+
 
 }
