+++ src/main/java/org/thingai/app/api/v1/RankApi.java
@@ -1,0 +1,10 @@
+package org.thingai.app.api.v1;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/rank")
+public class RankApi {
+
+}
