+++ src/main/java/org/thingai/app/api/v1/TeamApi.java
@@ -1,0 +1,9 @@
+package org.thingai.app.api.v1;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/team")
+public class TeamApi {
+}
